Week 2/Day 1/App/requirements.txt
fastapi==0.85.0
uvicorn==0.20.0
PyPDF2==3.0.1  # Updated to match your installed version
wandb==0.17.1  # Updated to match your installed version
scikit-learn==1.5.0  # Updated to match your installed version
python-multipart==0.0.5
python-dotenv==1.0.1  # Updated to match your installed version